feat: update logo urls, sync restaurant detail and category screen with firestore, add banner detail with copyable vouchers

+++ app/src/main/java/com/example/a43_kltn_ttfood/data/model/FirestoreModels.kt
@@ -281,6 +281,6 @@ data class CartItem(
 ) {
     @get:Exclude
     val foodBgColor: androidx.compose.ui.graphics.Color
-        get() = androidx.compose.ui.graphics.Color(foodBgColorVal.toULong())
+        get() = androidx.compose.ui.graphics.Color(foodBgColorVal.toInt())
 }
